Aufgabe 5.2 (WIP)

+++ framework/shape.hpp
@@ -3,8 +3,8 @@
 
 class Shape {
   public: 
-    virtual double area() const = 0;
-    virtual double volume() const = 0;
+    virtual float area() const = 0;
+    virtual float volume() const = 0;
 }
 
 #endif
